requirements.txt
fastapi
uvicorn[standard]
vosk
pydub
python-multipart
aiofiles
pytest
httpx
pytest-asyncio
openai-whisper